test(textarea): migrate states
