+++ core/src/components/textarea/test/states/textarea.e2e.ts-snapshots/textarea-disabled-ios-ltr-Mobile-Chrome-linux.png
