+++ core/src/components/textarea/test/states/textarea.e2e.ts-snapshots/textarea-readonly-ios-ltr-Mobile-Safari-linux.png
