+++ core/src/components/textarea/test/states/textarea.e2e.ts-snapshots/textarea-readonly-md-ltr-Mobile-Firefox-linux.png
